adding seeds

+++ db/schema.sql
@@ -12,7 +12,7 @@ CREATE TABLE department (
 CREATE TABLE role (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(30),
-    salary INT,
+    salary DECIMAL,
     departmentId INT,
     FOREIGN KEY (departmentId) REFERENCES department(id)
 );
